Rename Access endpoints to Public endpoints
